Fixed image files getting filtered instead of ignored

+++ main.py
@@ -38,44 +38,50 @@ while True:
     elif event == 'emptyList': # if user empties list box(list of files)
             window['LB'].update([])
             LB_vals = []
-    elif event == 'inputFiles': # if user adds files
+    elif event == 'inputFiles': # if user adds files (add files button)
             input = values['inputFiles'].split(';') # split the list of inputted files at the semicolon
             newFiles = []
             for i in input: # add filepaths to newFiles list
                 newFiles.append(i)
-            window.find_element('LB').Update(newFiles) # display listbox values
+            window.find_element('LB').Update(newFiles) # update the window with listbox values
             LB_vals = newFiles # list for listbox values
 
     elif event == 'run': # if user hits run button
         listBox = LB_vals
-        delete = False # variable for whether delete mode is enable or disable
+        delete = False # variable for whether delete mode is enabled or disabled
         worked = False # variable for if operation succeeded or failed
+        supportedExtensions = ['.webm','.mpg','.mp2','.mpeg','.mpe','.mpv',
+                               '.ogg','.mp4','.m4p','.m4v','.m4v','.avi',
+                               '.wmv','.mov','.qt','.flv','.swf','.hevc',
+                               '.heic']
         
-        # TODO: add support for ignoring images
         if listBox:
             for file in listBox:
-                if values['delete'] == True:
-                    delete = True # enables delete mode
+                fileAndExt = os.path.splitext(file)
+                extension = fileAndExt[1].lower()
+                if (extension in supportedExtensions):
+                    if values['delete'] == True:
+                        delete = True # enables delete mode
+                        
+                    # get video resolution
+                    vid = cv2.VideoCapture(file)
+                    width = vid.get(cv2.CAP_PROP_FRAME_WIDTH)
+                    height = vid.get(cv2.CAP_PROP_FRAME_HEIGHT)
                     
-                # get video resolution
-                vid = cv2.VideoCapture(file)
-                width = vid.get(cv2.CAP_PROP_FRAME_WIDTH)
-                height = vid.get(cv2.CAP_PROP_FRAME_HEIGHT)
-                
-                # get target resolution
-                targetWidth = values[0]
-                targetHeight = values[1]
-                
-                # operate on files
-                try:
-                    if width == targetWidth and height == targetHeight:
-                        if delete:
-                            os.remove(file)
-                        else:
-                            shutil.move(file, values['outputFolder'])
-                    worked = True
-                except:
-                    sg.Popup('Error Occurred', keep_on_top=True)
+                    # get target resolution
+                    targetWidth = values[0]
+                    targetHeight = values[1]
+                    
+                    # operate on files
+                    try:
+                        if width == targetWidth and height == targetHeight:
+                            if delete: # if delete mode is enable, delete it
+                                os.remove(file)
+                            else: # otherwise move it to the output folder
+                                shutil.move(file, values['outputFolder'])
+                        worked = True
+                    except:
+                        sg.Popup('Error Occurred', keep_on_top=True)
             if worked:
                 sg.Popup('Operation succeeded', keep_on_top = True)
